# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Tim Burton Movies",
      "version": "1.0.0",
      "description": "API Movies Tim Burton"
    },
    "servers": [
      {
        "url": "https:/",
        "description": "API de peliculas de Tim Burton. Cada tarjeta de una pelicula, contiene el titulo de la pelicula, año de estreno, genero, director,synopsis y una portada imagen."
      }
    ],
    "paths": {
      "/": {
        "get": {
          "summary": "Landing con galeria de todos las peliculas",
          "responses": {
            "200": {
              "description": "Operación exitosa",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean"
                      },
                      "data": {
                        "type": "array",
                        "items": {
                       
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/register": {
        "post": {
          "summary": "Registrar un usuario",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                
              }
            }
          },
          "responses": {
            "200": {
              "description": "Registro exitoso"
            },
            "404": {
              "description": "El usuario ya esta registrado"
            }
          }
        }
      },
      "/login": {
        "post": {
          "summary": "login usuario",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                
              }
            }
          },
          "responses": {
            "200": {
              "description": "Todo ha salido bien"
            },
            "404": {
              "description": "contraseña incorrecta"
            }
          }
        }
      },
      "/profileUser": {
        "get": {
          "summary": "autenticacion",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                
              }
            }
          },
          "responses": {
            "200": {
              "description": "Todo ha salido bien"
            },
            "500": {
              "description": "error al procesar la solicitud"
            }
          }
        }
      },
      "/logout": {
        "post": {
          "summary": "Cierre sesion usuario",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                
              }
            }
          },
          "responses": {
            "200": {
              "description": "sesion cerrada exitosamente"
            },
            "404": {
              "description": "Ha ocurrido un error, intentalo denuevo"
            }
          }
        }
      },
      "/movies": {
        "post": {
          "summary": "Crea una nueva pelicula",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                
              }
            }
          },
          "responses": {
            "200": {
              "description": "Pelicula creado exitosamente"
            },
            "404": {
              "description": "La pelicula ya existe"
            }
          }
        }
      },
      "/movies/name/:name": {
        "get": {
          "summary": "Busca una pelicula por nombre",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "exito"
            },
            "404": {
              "description": "No se encontro ninguna pelicula con ese nombre"
            }
          }
        }
      },
      "movies/:id": {
        "put": {
          "summary": "Hacer un cambio en una pelicula (actualizar)",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Pelicula actualizada exitosamente"
            },
            "404": {
              "description": "La Pelicula no se ha encontrado"
            }
          }
        }
      },
      "/movies/": {
        "Delete": {
          "summary": "Elimina una pelicula por nombre",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                
              }
            }
          },
          "responses": {
            "200": {
              "description": "Pelicula eliminada exitosamente"
            },
            "404": {
              "description": "No se ha eliminado, ya que no existe"
            }
          }
        }
      },
      "/movies/:id": {
        "get": {
          "summary": "Obtiene detalles de una pelicula específica",
          "parameters": [
            {
              "name": "idMovie",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Operación exitosa"
            },
            "404": {
              "description": "El id no existe"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Movies": {
          "type": "object",
          "properties": {
            "idMovie": {
              "type": "number"
            },
            "name": {
              "type": "string"
            },
            "year": {
              "type": "number"
            },
            "gender": {
              "type": "string"
            },
            "director": {
              "type": "string"
            },
            "synopsis": {
              "type": "string"
            },
            "image": {
              "type": "string"
            }
        
            }
           
          }
        },
        "NewMovie": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "year": {
              "type": "number"
            },
            "gender": {
              "type": "string"
            },
            "director": {
              "type": "string"
            },
            "synopsis": {
              "type": "string"
            },
            
            "image": {
              "type": "string"
            }
          }
        }
      }
    }
  
  
  